requirements.txt
pandas
numpy
requests
apache-airflow
apache-airflow-providers-amazon
apache-airflow-providers-postgres